[LibDroneMovement] Add AtCommandFactory Class for the libDroneMovement
Our project needs an embedded version of the ARDrone Middleware
in order to let the drone be able to move by himself.
The SDK seems too complex to this job.

The AtCommandFactory class is the first part of this Middleware.
It allows to create drone's AtCommands and handle the sequence
number autonomously.

It's also the first part of the libDroneMovement which will be
embedded in the drone.

Signed-off-by: clero <clero@etud.insa-toulouse.fr>