[LibDroneVideo] Enhance line detection algorithm.
This patch introduces the LineDetector object which can be used to
analyse a frame and to find a line on it. This file uses the algorithm
developed in the LineDetection class. It adds some good object features
and we can now tell to the drone the line position and if he is flying
on the right way.

Signed-off-by: clero <clero@etud.insa-toulouse.fr>