[LibDroneVideo] Add LineDetectionError exception
When the drone is flying, the line can often be lost or
badly analysed.
This patch introduces the LineDetectionError exception
which is raised when LineDetector internal algorithm
detects something strange in analysed data.
For example, when the line angle is NaN or when the algorithm
deduces that the drone needs to go left and right at the same
time.

Signed-off-by: clero <clero@etud.insa-toulouse.fr>