[Core] Implement Mission step's creator classes
In order to use easily mission steps, we need to implement the
command design pattern. This patch introduces factory and parser
for mission steps. Those class allows to create mission steps
from a json files. Those steps will be stored in a queue in order
to launch action implemented in specialized child classes.

Signed-off-by: clero <clero@etud.insa-toulouse.fr>