[libDroneMovement] Adapt drone AT Command sending timing
This patch introduces a basic algorithm whixh allow to get smoother
movement when sending AT command to the drone. Previously, we were
sending an hover command as soon as the sending queue was empty.
Now, we poll rapidly the queue during 30ms before sending the
hovering command. It allows to be sure that we are not missing
a command of a producer just because of a producer desynchronisation.

Signed-off-by: clero <clero@etud.insa-toulouse.fr>